:sparkles: (floating_point) no pipeline forced

+++ floating_point/src/testbench.cc
@@ -5,14 +5,14 @@
 int main() {
     srand(125);
     for (unsigned int itest = 0, ntest = 20; itest <= ntest; ++itest) {
-        double a = static_cast<double>(rand() & 0xFFF);
-        double b = static_cast<double>(rand() & 0xFF);
+        float a = static_cast<float>(rand() & 0xFFF);
+        float b = static_cast<float>(rand() & 0xFF);
 
         printf("a = %f, b = %f, ", a, b);
         bool ok = true;
 
-        double sum_exp = a + b;
-        double sum_fw = sum(a, b);
+        float sum_exp = a + b;
+        float sum_fw = sum(a, b);
         printf(" sum = %f (expected %f), %s\n", sum_fw, sum_exp, sum_fw == sum_exp ? " ok " : "fail");
         ok = ok && (sum_fw == sum_exp);
